ImageHttpReader: use caller-provided HttpClient
Since HttpClient is now stateful, use a caller-provided client to
request the image. Since ImageHttpReader lives for only a single image
request, for now the caller is responsible for maintaining the lifetime
of the HttpClient longer than the lifetime of the reader, although this
can easily be changed to a shared pointer if non-encompassing lifetimes
are necessary.