ConnectedComponents: make value simple bytes array
Since value may be many different types and is otherwise not important
to the behavior of connected components, just store it as a byte array
to be compatible with label types. This was chosen over templating,
since templating CC would make using it cumbersome for most consumers,
that do not care about the value type.